modularize package structure by removing 'webapi' prefix from adapter classes
