+++ application/src/main/java/org/mandarin/booking/app/AuthService.java
@@ -3,7 +3,7 @@ package org.mandarin.booking.app;
 import lombok.RequiredArgsConstructor;
 import org.mandarin.booking.AuthException;
 import org.mandarin.booking.TokenHolder;
-import org.mandarin.booking.adapter.security.TokenUtils;
+import org.mandarin.booking.adapter.TokenUtils;
 import org.mandarin.booking.app.persist.MemberQueryRepository;
 import org.mandarin.booking.app.port.AuthUseCase;
 import org.mandarin.booking.domain.member.Member;
