+++ application/src/main/java/org/mandarin/booking/app/CustomAuthenticationProvider.java
@@ -2,7 +2,7 @@ package org.mandarin.booking.app;
 
 import lombok.RequiredArgsConstructor;
 import org.mandarin.booking.AuthException;
-import org.mandarin.booking.adapter.security.CustomMemberAuthenticationToken;
+import org.mandarin.booking.adapter.CustomMemberAuthenticationToken;
 import org.mandarin.booking.app.persist.MemberQueryRepository;
 import org.mandarin.booking.domain.member.Member;
 import org.springframework.security.authentication.AuthenticationProvider;
