+++ application/src/main/java/org/mandarin/booking/app/JwtTokenUtils.java
@@ -18,7 +18,7 @@ import org.jspecify.annotations.Nullable;
 import org.mandarin.booking.AuthException;
 import org.mandarin.booking.MemberAuthority;
 import org.mandarin.booking.TokenHolder;
-import org.mandarin.booking.adapter.security.TokenUtils;
+import org.mandarin.booking.adapter.TokenUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
