+++ application/src/test/java/org/mandarin/booking/adapter/security/CustomAuthenticationProviderTest.java
@@ -6,6 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 import org.mandarin.booking.IntegrationTest;
+import org.mandarin.booking.adapter.CustomMemberAuthenticationToken;
 import org.mandarin.booking.app.CustomAuthenticationProvider;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
