+++ application/src/test/java/org/mandarin/booking/adapter/security/JwtFilterTest.java
@@ -1,15 +1,17 @@
 package org.mandarin.booking.adapter.security;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.mandarin.booking.adapter.webapi.ApiStatus.FORBIDDEN;
-import static org.mandarin.booking.adapter.webapi.ApiStatus.SUCCESS;
-import static org.mandarin.booking.adapter.webapi.ApiStatus.UNAUTHORIZED;
+import static org.mandarin.booking.adapter.ApiStatus.FORBIDDEN;
+import static org.mandarin.booking.adapter.ApiStatus.SUCCESS;
+import static org.mandarin.booking.adapter.ApiStatus.UNAUTHORIZED;
 
 import java.util.List;
 import org.junit.jupiter.api.Test;
 import org.mandarin.booking.IntegrationTest;
 import org.mandarin.booking.IntegrationTestUtils;
 import org.mandarin.booking.NoRestDocs;
+import org.mandarin.booking.adapter.JwtFilter;
+import org.mandarin.booking.adapter.TokenUtils;
 import org.mandarin.booking.adapter.security.JwtFilterTest.TestAuthController;
 import org.mandarin.booking.adapter.security.JwtFilterTest.TestAuthController.TestSecurityConfig;
 import org.springframework.beans.factory.annotation.Autowired;
