+++ application/src/test/java/org/mandarin/booking/webapi/not_found/GET_specs.java
@@ -1,7 +1,7 @@
-package org.mandarin.booking.adapter.webapi;
+package org.mandarin.booking.webapi.not_found;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.mandarin.booking.adapter.webapi.ApiStatus.NOT_FOUND;
+import static org.mandarin.booking.adapter.ApiStatus.NOT_FOUND;
 
 import org.junit.jupiter.api.Test;
 import org.mandarin.booking.IntegrationTest;
@@ -11,10 +11,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 
 @IntegrationTest
 @NoRestDocs
-class GlobalExceptionHandlerTest {
-
+public class GET_specs {
     @Test
-    void endpointNotFound(@Autowired IntegrationTestUtils testUtils){
+    void endpointNotFound(@Autowired IntegrationTestUtils testUtils) {
         // Act
         var request = testUtils.get("/not-found")
                 .assertFailure();
