+++ application/src/test/java/org/mandarin/booking/webapi/show/POST_specs.java
@@ -2,10 +2,10 @@ package org.mandarin.booking.webapi.show;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mandarin.booking.MemberAuthority.ADMIN;
-import static org.mandarin.booking.adapter.webapi.ApiStatus.BAD_REQUEST;
-import static org.mandarin.booking.adapter.webapi.ApiStatus.INTERNAL_SERVER_ERROR;
-import static org.mandarin.booking.adapter.webapi.ApiStatus.SUCCESS;
-import static org.mandarin.booking.adapter.webapi.ApiStatus.UNAUTHORIZED;
+import static org.mandarin.booking.adapter.ApiStatus.BAD_REQUEST;
+import static org.mandarin.booking.adapter.ApiStatus.INTERNAL_SERVER_ERROR;
+import static org.mandarin.booking.adapter.ApiStatus.SUCCESS;
+import static org.mandarin.booking.adapter.ApiStatus.UNAUTHORIZED;
 
 import java.time.LocalDate;
 import java.util.List;
