+++ internal/src/main/java/org/mandarin/booking/adapter/ApiResponse.java
@@ -1,4 +1,4 @@
-package org.mandarin.booking.adapter.webapi;
+package org.mandarin.booking.adapter;
 
 import java.time.LocalDateTime;
 import lombok.Getter;
