+++ internal/src/main/java/org/mandarin/booking/adapter/ApiStatus.java
@@ -1,0 +1,10 @@
+package org.mandarin.booking.adapter;
+
+public enum ApiStatus {
+    SUCCESS,
+    BAD_REQUEST,
+    UNAUTHORIZED,
+    INTERNAL_SERVER_ERROR,
+    FORBIDDEN,
+    NOT_FOUND
+}
