+++ internal/src/main/java/org/mandarin/booking/adapter/CustomAccessDeniedHandler.java
@@ -1,6 +1,6 @@
-package org.mandarin.booking.adapter.security;
+package org.mandarin.booking.adapter;
 
-import static org.mandarin.booking.adapter.webapi.ApiStatus.FORBIDDEN;
+import static org.mandarin.booking.adapter.ApiStatus.FORBIDDEN;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.HttpServletRequest;
@@ -8,7 +8,6 @@ import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.mandarin.booking.adapter.webapi.ErrorResponse;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.web.access.AccessDeniedHandler;
 import org.springframework.stereotype.Component;
@@ -16,7 +15,7 @@ import org.springframework.stereotype.Component;
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class CustomAccessDeniedHandler implements AccessDeniedHandler {
+class CustomAccessDeniedHandler implements AccessDeniedHandler {
     private final ObjectMapper objectMapper;
 
     @Override
