+++ internal/src/main/java/org/mandarin/booking/adapter/CustomAuthenticationEntryPoint.java
@@ -1,19 +1,17 @@
-package org.mandarin.booking.adapter.security;
+package org.mandarin.booking.adapter;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import lombok.RequiredArgsConstructor;
-import org.mandarin.booking.adapter.webapi.ApiStatus;
-import org.mandarin.booking.adapter.webapi.ErrorResponse;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.AuthenticationEntryPoint;
 import org.springframework.stereotype.Component;
 
 @Component
 @RequiredArgsConstructor
-public class CustomAuthenticationEntryPoint implements AuthenticationEntryPoint {
+class CustomAuthenticationEntryPoint implements AuthenticationEntryPoint {
     private final ObjectMapper objectMapper;
 
     @Override
