+++ internal/src/main/java/org/mandarin/booking/adapter/JwtFilter.java
@@ -1,4 +1,4 @@
-package org.mandarin.booking.adapter.security;
+package org.mandarin.booking.adapter;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
