+++ internal/src/main/java/org/mandarin/booking/adapter/package-info.java
@@ -1,4 +1,4 @@
 @NullMarked
-package org.mandarin.booking.adapter.security;
+package org.mandarin.booking.adapter;
 
 import org.jspecify.annotations.NullMarked;
