--- internal/src/main/java/org/mandarin/booking/adapter/webapi/package-info.java
@@ -1,4 +1,0 @@
-@NullMarked
-package org.mandarin.booking.adapter.webapi;
-
-import org.jspecify.annotations.NullMarked;
